+++ tapis-appslib/src/main/java/edu/utexas/tacc/tapis/apps/service/AppsService.java
@@ -56,7 +56,7 @@ public interface AppsService
   boolean isEnabled(ResourceRequestUser rUser, String appId)
           throws TapisException, TapisClientException, NotAuthorizedException;
 
-  App getApp(ResourceRequestUser rUser, String appId, String appVersion, boolean requireExecPerm, boolean skipTapisAuth)
+  App getApp(ResourceRequestUser rUser, String appId, String appVersion, boolean requireExecPerm, String impersonationId)
           throws TapisException, TapisClientException, NotAuthorizedException;
 
   int getAppsTotalCount(ResourceRequestUser rUser, List<String> searchList, List<OrderBy> orderByList,