+++ CHANGELOG.md
@@ -9,6 +9,18 @@ You may also reference live-docs based on the openapi specification here:
 https://tapis-project.github.io/live-docs
 
 ---------------------------------------------------------------------------
+## 1.1.3 - 2022-??-??
+
+Incremental improvements and preview of new features.
+
+### New features:
+- Refactor authorization checks for maintainability.
+- Replace skipTapisAuthorization with impersonationId for requests from Jobs service.
+
+### Bug fixes:
+- None
+
+---------------------------------------------------------------------------
 ## 1.1.2 - 2022-04-13
 
 Preview of new features.